correcoes

+++ bankserver/src/main/java/com/bankserver/dto/request/ClienteRegistrationDTO.java
@@ -1,10 +1,12 @@
 package com.bankserver.dto.request;
 
+import java.math.BigDecimal;
+
 public record ClienteRegistrationDTO(
                 @NotBlank String cpf,
                 @NotBlank String email,
                 @NotBlank String nome,
                 @NotBlank String telefone,
-                Double salario,
+                BigDecimal salario,
                 @Valid EnderecoDTO endereco) {
 }
