+++ bankserver/src/main/java/com/bankserver/dto/response/R09ResDTO.java
@@ -1,8 +1,10 @@
 package com.bankserver.dto.response;
 
+import java.math.BigDecimal;
+
 public record R09ResDTO(
         Long contaId,
         String cpf,
         String nome,
-        Double salario) {
+        BigDecimal salario) {
 }
